Fixes the analog sensor issues with the arduino code
There were two major problems. One was that I had signed chars
instead of unsigned chars that I was sending back as the analog
data. This sometimes caused the code to send back negative values.
This may not have actually been an issue, because unsigned chars
and signed chars are actually the same things and pythong may have
interpretted the data correctly.

The biggest problem was that we had an issue before, where the
firmware seemed to sometimes send back empty data packets (just a
newline), so we had code in the python to throw this out. It turns
out that this problem had since been fixed, but any char that we
read in that matched LF or CR was being thrown out. Thus whenever
the analog data happened to be one of those values, it was thrown
out, and the whole data packet was thrown off by a character,
meaning python would hang, expecting more and not getting it.